+++ outputs/all/all.go
@@ -5,6 +5,8 @@ import (
 	_ "github.com/influxdb/telegraf/outputs/datadog"
 	_ "github.com/influxdb/telegraf/outputs/influxdb"
 	_ "github.com/influxdb/telegraf/outputs/kafka"
+	_ "github.com/influxdb/telegraf/outputs/librato"
 	_ "github.com/influxdb/telegraf/outputs/mqtt"
+	_ "github.com/influxdb/telegraf/outputs/nsq"
 	_ "github.com/influxdb/telegraf/outputs/opentsdb"
 )